+++ src/app/components/footer/contact-form/contact-form-message-and-button/styles.tsx
@@ -1,0 +1,22 @@
+export const footerContactUsFormMessageAndButtonBoxStyles = {
+    display: "flex",
+    flexDirection: "column",
+    height: "300px",
+};
+
+export const footerContactUsFormLabelStyles = {
+    marginBottom: "6px",
+};
+
+export const footerContactUsFormMessageTextFieldStyles = {
+    marginBottom: "20px",
+};
+
+export const footerContactUsFormButtonStyles = {
+    backgroundColor: "rgb(198, 164, 126)",
+    color: "#2b2b2b",
+    alignSelf: "flex-end",
+    "&:hover": {
+        backgroundColor: "rgb(198, 164, 126)",
+    },
+};
