+++ src/app/components/footer/contact-form/contact-form-name-and-email/styles.tsx
@@ -1,0 +1,10 @@
+export const footerContactUsFormNameAndEmailBoxStyles = {
+    display: "flex",
+    flexDirection: "column",
+    marginRight: "20px",
+    height: "200px",
+};
+
+export const footerContactUsFormLabelStyles = {
+    marginBottom: "6px",
+};
