+++ src/app/components/footer/contact-form/styles.tsx
@@ -19,25 +19,7 @@ export const footerContactUsFormNameAndEmailBoxStyles = {
     height: "200px",
 };
 
-export const footerContactUsFormMessageAndButtonBoxStyles = {
-    display: "flex",
-    flexDirection: "column",
-    height: "300px",
-};
-
 export const footerContactUsFormLabelStyles = {
     marginBottom: "6px",
 };
 
-export const footerContactUsFormButtonStyles = {
-    backgroundColor: "rgb(198, 164, 126)",
-    color: "#2b2b2b",
-    alignSelf: "flex-end",
-    "&:hover": {
-        backgroundColor: "rgb(198, 164, 126)",
-    },
-};
-
-export const footerContactUsFormMessageTextFieldStyles = {
-    marginBottom: "20px",
-};
